avto/requirements.txt
Django==4.0.2
psycopg2==2.9.3
djangorestframework==3.13.1
django-debug-toolbar==3.2.3
Pillow==8.4.0
django-filter==21.1
django-simple-captcha==0.5.14
pycodestyle==2.8.0